tests: add riscv clic qtest case and a function in qtest
This adds riscv32-clic-test.c, containing qtest test cases for configuring
CLIC (via virt machine) and for triggering interrupts.

In order to detect the interrupts, qtest.c has been updated to send interrupt
information back to the test about the IRQ being delivered. Since we need to
both trigger and detect the interrupt, qtest has also been updated to allow
both an input and an output GPIO to be intercepted.

Signed-off-by: Troy Song <wb-szh830103@alibaba-inc.com>
Signed-off-by: Ian Brockbank <ian.brockbank@cirrus.com>